Organised edit routine related views into one package
