--- src/view/EditRoutineView.java
@@ -1,4 +1,0 @@
-package view;
-
-public class EditRoutineView {
-}
