+++ src/view/edit_routine/AddExerciseView.java
@@ -1,4 +1,4 @@
-package view;
+package view.edit_routine;
 
 public class AddExerciseView {
 }
