+++ src/view/edit_routine/DeleteExerciseView.java
@@ -1,4 +1,4 @@
-package view;
+package view.edit_routine;
 
 public class DeleteExerciseView {
 }
