+++ src/view/edit_routine/ReorderRoutineView.java
@@ -1,0 +1,6 @@
+package view.edit_routine;
+
+public class ReorderRoutineView {
+
+    public final String viewName = "edit";
+}
